# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_goodreads": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_goodreads",
        "type": "object"
      },
...
      "BookOut": {
        "properties": {
          "active_loan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoanOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "added_at": {
            "format": "date-time",
            "title": "Added At",
            "type": "string"
          },
          "added_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "categories": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "google_books_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Books Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_private": {
            "default": false,
            "title": "Is Private",
            "type": "boolean"
          },
          "isbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "my_finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Finished At"
          },
          "my_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Rating"
          },
          "my_started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Started At"
          },
          "my_status": {
            "$ref": "#/components/schemas/ReadStatus",
            "default": "unread"
          },
          "ownership": {
            "$ref": "#/components/schemas/OwnershipStatus",
            "default": "owned"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "series_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Index"
          },
          "series_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Name"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "tags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "isbn",
          "title",
          "subtitle",
          "author",
          "publisher",
          "year",
          "description",
          "cover_url",
          "added_at"
        ],
        "title": "BookOut",
        "type": "object"
      },
...
      "GoodreadsImportOut": {
        "description": "What an import actually did.\n\nReported field by field rather than as a single count, because the useful\nquestion after an import is usually \"why did it not pick up book X?\" and\nthe answer is almost always one of: the shelf was not one we map, the book\nis not in the catalogue, or the status was already correct.",
        "properties": {
          "created": {
            "description": "Books added from the export",
            "minimum": 0.0,
            "title": "Created",
            "type": "integer"
          },
          "matched": {
            "description": "Rows matched to a book already here",
            "minimum": 0.0,
            "title": "Matched",
            "type": "integer"
          },
          "rows_read": {
            "description": "Rows with a shelf we understand",
            "minimum": 0.0,
            "title": "Rows Read",
            "type": "integer"
          },
          "skipped": {
            "description": "Rows on a shelf we do not map",
            "minimum": 0.0,
            "title": "Skipped",
            "type": "integer"
          },
          "statuses_updated": {
            "description": "Statuses actually changed",
            "minimum": 0.0,
            "title": "Statuses Updated",
            "type": "integer"
          },
          "unmatched_titles": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Unmatched Titles",
            "type": "array"
          }
        },
        "required": [
          "rows_read",
          "matched",
          "created",
          "statuses_updated",
          "skipped"
        ],
        "title": "GoodreadsImportOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoanCreate": {
        "properties": {
          "book_id": {
            "title": "Book Id",
            "type": "integer"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "loaned_to_user_id": {
            "title": "Loaned To User Id",
            "type": "integer"
          }
        },
        "required": [
          "book_id",
          "loaned_to_user_id"
        ],
        "title": "LoanCreate",
        "type": "object"
      },
      "LoanOut": {
        "properties": {
          "book": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "book_id": {
            "title": "Book Id",
            "type": "integer"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_overdue": {
            "default": false,
            "title": "Is Overdue",
            "type": "boolean"
          },
          "loaned_at": {
            "format": "date-time",
            "title": "Loaned At",
            "type": "string"
          },
          "loaned_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "loaned_by_user_id": {
            "title": "Loaned By User Id",
            "type": "integer"
          },
          "loaned_to": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "loaned_to_user_id": {
            "title": "Loaned To User Id",
            "type": "integer"
          },
          "returned_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned At"
          }
        },
        "required": [
          "id",
          "book_id",
          "loaned_to_user_id",
          "loaned_by_user_id",
          "loaned_at",
          "returned_at"
        ],
        "title": "LoanOut",
        "type": "object"
      },
...
      "OwnershipStatus": {
        "description": "Whether a book is physically on the shelf.\n\nDeliberately separate from ReadStatus, which is about a person, not an\nobject. \"I have read this\" and \"we own a copy of this\" are independent\nclaims: a library borrowing is read but not owned, and an unread gift is\nowned but not read. Conflating them is what makes an imported reading\nhistory look like a catalogue of possessions.\n\nUNKNOWN exists because a Goodreads export cannot answer the question at\nall. Defaulting those rows to OWNED would assert something nobody checked;\ndefaulting them to NOT_OWNED would be an equally unfounded guess. So they\narrive unverified and wait to be confirmed.",
        "enum": [
          "owned",
          "not_owned",
          "unknown"
        ],
        "title": "OwnershipStatus",
        "type": "string"
      },
...
      "Page_LoanOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LoanOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[LoanOut]",
        "type": "object"
      },
...
      "ReadStatus": {
        "description": "A member's own progress through a book. Stored in `user_books.status`.\n\nWANT_TO_READ is deliberately distinct from UNREAD: \"on the shelf, not\nstarted\" and \"I intend to read this\" are different statements, and\nGoodreads exports carry the distinction (its `to-read` shelf), so\ncollapsing them would lose information on import.",
        "enum": [
          "unread",
          "want_to_read",
          "reading",
          "read"
        ],
        "title": "ReadStatus",
        "type": "string"
      },
...
      "TagCategory": {
        "description": "The three groups tags are presented in throughout the UI.",
        "enum": [
          "type",
          "genre",
          "age"
        ],
        "title": "TagCategory",
        "type": "string"
      },
      "TagOut": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/TagCategory"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "category"
        ],
        "title": "TagOut",
        "type": "object"
      },
...
      "UserOut": {
        "description": "A member as seen by other members. Deliberately has no password field.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "is_admin",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Catalogue, lend and track a family's physical book collection.",
    "title": "Endpaper",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/imports/goodreads": {
      "post": {
        "description": "Apply the reading statuses from a Goodreads export.\n\nStatuses are **personal**, so this only ever writes the importing member's\nown `user_books` rows. Importing your shelves does not change what anyone\nelse has read, and two members can import their own exports without\nfighting over the same books.\n\nBooks created by `create_missing` are marked `ownership=unknown`: a reading\nhistory is not evidence of possession. They can then be confirmed together\nfrom the library view, which is what the bulk ownership endpoint is for.",
        "operationId": "import_goodreads",
        "parameters": [
          {
            "description": "Add books from the export that are not in the catalogue",
            "in": "query",
            "name": "create_missing",
            "required": false,
            "schema": {
              "default": false,
              "description": "Add books from the export that are not in the catalogue",
              "title": "Create Missing",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_goodreads"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodreadsImportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Import Goodreads",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/loans": {
      "get": {
        "operationId": "list_loans",
        "parameters": [
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "overdue_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Overdue Only",
              "type": "boolean"
            }
          },
          {
            "description": "1-based page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Rows per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Rows per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_LoanOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Loans",
        "tags": [
          "loans"
        ]
      },
      "post": {
        "operationId": "create_loan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Loan",
        "tags": [
          "loans"
        ]
      }
    },
    "/api/loans/{loan_id}/return": {
      "put": {
        "description": "Recording a return is a shelf action, not an ownership one, so any member\nmay do it, for any book they can see.",
        "operationId": "return_loan",
        "parameters": [
          {
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "title": "Loan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Return Loan",
        "tags": [
          "loans"
        ]
      }
    },
...
  }
}